Refactor folders & set bg color
