--- bookstore-client/src/components/admin-panel/ManageRequests.vue
@@ -1,3 +1,0 @@
-<template>
-    <h1>Leenverzoeken</h1>
-</template>
