+++ bookstore-client/src/components/admin-panel/book/DeleteBook.vue
