+++ bookstore-client/src/components/details/BookDetail.vue
