+++ bookstore-client/src/components/details/ViewPerson.vue
