+++ bookstore-client/src/router/index.js
@@ -1,46 +1,35 @@
 import { createRouter, createWebHashHistory } from 'vue-router'
-import HomeView from '../views/HomeView.vue'
-import AdminView from '../views/admin-panel/AdminView.vue'
-import EditBooksView from '../views/admin-panel/EditBooksView.vue'
+// import HomeView from '../views/HomeView.vue'
+// import AdminView from '../views/admin-panel/AdminView.vue'
+import EditBooksView from '../views/admin-panel/book/EditBooksView.vue'
 import RequestsView from '../views/admin-panel/RequestsView.vue'
-import EditUsersView from '../views/admin-panel/EditUsersView.vue'
-import AddUserView from '../views/admin-panel/AddUserView.vue'
-import AddBookView from '../views/admin-panel/AddBookView.vue'
+import EditUsersView from '../views/admin-panel/user/EditUsersView.vue'
+import AddUserView from '../views/admin-panel/user/AddUserView.vue'
+import AddBookView from '../views/admin-panel/book/AddBookView.vue'
 import AddCopyView from '../views/admin-panel/AddCopyView.vue'
-import LoanedBooksView from '../views/admin-panel/LoanedBooksView.vue'
+import LoanedBooksView from '../views/admin-panel/book/LoanedBooksView.vue'
 import ReservationsView from '../views/admin-panel/ReservationsView.vue'
-import LoginView from '../views/LoginView.vue'
+import LoginView from '../views/general-pages/LoginView.vue'
 import MyBooksView from '../views/MyBooksView.vue'
-import ContactView from '../views/ContactView.vue'
-import MyTestView from '../views/MyTestView.vue'
-import MyAccountView from '../views/MyAccountView.vue'
+import ContactView from '../views/general-pages/ContactView.vue'
+import MyTestView from '../views/test/MyTestView.vue'
+import MyAccountView from '../views/general-pages/MyAccountView.vue'
 import CatalogueView from '../views/CatalogueView.vue'
-import BookDetailView from '../views/BookDetailView.vue'
-import UpdateBookView from '../views/admin-panel/UpdateBookView.vue'
-import CreateLoanFromReservation from '../components/admin-panel/CreateLoanFromReservation'
-
-// Endpoints
-// “/” 
-// “/login/”
-// “/mybooks/
-// “/book/$`id`”
-// “/contact”
-// “/admin/requests
-// “/admin/edit-users” 
-// “/admin/edit-books”
-// “/admin/add-user”
+import BookDetailView from '../views/detail-pages/BookDetailView.vue'
+import UpdateBookView from '../views/admin-panel/book/UpdateBookView.vue'
+import CreateLoanFromReservation from '../components/admin-panel/reservation-overview/CreateLoanFromReservation'
 
 const routes = [
-  {
-    path: '/',
-    name: 'home',
-    component: HomeView
-  },
-  {
-    path: '/admin',
-    name: 'admin',
-    component: AdminView
-  },
+  // {
+  //   path: '/',
+  //   name: 'home',
+  //   component: HomeView
+  // },
+  // {
+  //   path: '/admin',
+  //   name: 'admin',
+  //   component: AdminView
+  // },
   {
     path: '/admin/requests',
     name: 'requests',
@@ -112,7 +101,7 @@ const routes = [
     component: MyTestView
   },
   {
-    path: '/catalogue',
+    path: '/',
     name: 'catalogue',
     component: CatalogueView
   },
