--- bookstore-client/src/views/HomeView.vue
@@ -1,21 +1,0 @@
-<template>
-  <div class="home">
-    <!-- <img alt="Vue logo" src="../assets/logo.png"> -->
-    <!-- <HelloWorld/> -->
-    <LoginCard></LoginCard>
-
-
-  </div>
-</template>
-
-<script>
-// @ is an alias to /src
-import LoginCard from '@/components/LoginCard.vue';
-
-export default {
-  name: 'HomeView',
-  components: {
-    LoginCard
-}
-}
-</script>
