+++ bookstore-client/src/views/admin-panel/AddCopyView.vue
@@ -7,8 +7,8 @@
   
   <script>
   // @ is an alias to /src
-  import AddCopy from '@/components/admin-panel/AddCopy.vue';
-  import SideBar from '@/components/SideBar.vue';
+  import AddCopy from '@/components/admin-panel/book/AddCopy.vue';
+  import SideBar from '@/components/reusable-components/SideBar.vue';
   
   export default {
     name: 'AddUserView',
