+++ bookstore-client/src/views/admin-panel/RequestsView.vue
@@ -8,8 +8,8 @@
   
 <script>
     // @ is an alias to /src
-    import CreateLoanFromReservation from '@/components/admin-panel/CreateLoanFromReservation.vue';
-    import SideBar from '@/components/SideBar.vue';
+    import CreateLoanFromReservation from '@/components/admin-panel/reservation-overview/CreateLoanFromReservation.vue';
+    import SideBar from '@/components/reusable-components/SideBar.vue';
     
     export default {
         name: 'RequestsView',
