+++ bookstore-client/src/views/general-pages/ContactView.vue
@@ -8,7 +8,7 @@
   
   <script>
   // @ is an alias to /src
-  import SideBar from '@/components/SideBar.vue';
+  import SideBar from '@/components/reusable-components/SideBar.vue';
   
   export default {
     name: 'ContactView',
